move PulsarMock to test, update snapshot test
